add the 'composition' package
Also adds required 'core' types to support the Composition framework, such
as Lazy<> and support for memoization of functions and properties.

The composition package supports both a decorated programming model and a
convention based programming model, both of which are modelled on the MEF
framework.

Abstract classes are not currently supported because TypeScript does not yet
support declarations for abstract classes ({new (...args: any[]): any}),
however, other approaches may be possible with some compiler trickery.