compile projects after installation
Because there are peer dependencies between projects and we can't resolve those
typings like we would for a normal dependency, we need to ensure that the typings
are made available to each subsequent project in the installation pipeline. We
can only do this by compiling the project after it has been installed.